refactor home chat flow into split UI and controller hook
