+++ components/home/search-params-handler.tsx
@@ -1,0 +1,21 @@
+"use client";
+
+import { useSearchParams } from "next/navigation";
+import { useEffect } from "react";
+
+export function SearchParamsHandler({ onReset }: { onReset: () => void }) {
+  const searchParams = useSearchParams();
+
+  useEffect(() => {
+    const reset = searchParams.get("reset");
+    if (reset === "true") {
+      onReset();
+
+      const newUrl = new URL(window.location.href);
+      newUrl.searchParams.delete("reset");
+      window.history.replaceState({}, "", newUrl.pathname);
+    }
+  }, [searchParams, onReset]);
+
+  return null;
+}
