+++ lib/prompt-suggestions.ts
@@ -1,0 +1,10 @@
+export const PROMPT_SUGGESTIONS = [
+  "Landing page",
+  "Todo app",
+  "Dashboard",
+  "Blog",
+  "E-commerce",
+  "Portfolio",
+  "Chat app",
+  "Calculator",
+] as const;
